Create README.md
I'm passionate about building software that makes a difference . Currently diving into Data Structures and Algorithms with java to sharpen my problem solving skills. I believe in continuous learning, whether it's mastering new programming languages, contributing to  open-source projects, or staying updated with the latest in tech. My goal is to combine my knowledge in backend and frontend development to create full-stack applications that are both efficient and user-friendly. Always open to collaboration and new opportunities to grow as a developer.